Fix compiler errors
Last number of changes were done away from the compiler, so this fixes those

+++ src/typedefs.h
@@ -5,8 +5,8 @@ using uword = std::uint16_t;
 using uint = std::uint32_t;
 using word = std::int16_t;
 
-using CpuClockCycles = std::int16_t;
-using PpuClockCycles = std::int16_t;
+using CpuClockCycles = std::int32_t;
+using PpuClockCycles = std::int32_t;
 
 //#define CpuToMaster( val ) (val * 3)
